requirements.txt
streamlit
pandas
numpy
scikit-learn
altair
google-generativeai
python-dotenv
reportlab
pyplot
